i18n, existing icon, move to utils, shadcn button
