+++ public/locale/en.json
@@ -1152,6 +1152,7 @@
   "insurance__policy_name": "Policy ID / Policy Name",
   "insurance_details_detail": "Include details of all the Insurance Policies held by the Patient for smooth insurance processing",
   "insurer_name_required": "Insurer Name is required",
+  "intended": "Intended",
   "intent": "Intent",
   "international_mobile": "International Mobile",
   "invalid_asset_id_msg": "Oops! The asset ID you entered does not appear to be valid.",
@@ -1436,6 +1437,7 @@
   "noshow": "No-show",
   "not_eligible": "Not Eligible",
   "not_specified": "Not Specified",
+  "not_taken": "Not Taken",
   "note": "Note",
   "notes": "Notes",
   "notes_placeholder": "Type your Notes",
@@ -1460,6 +1462,7 @@
   "old_password": "Current Password",
   "on": "on",
   "on_emergency_basis": " on emergency basis",
+  "on_hold": "On Hold",
   "ongoing_medications": "Ongoing Medications",
   "only_indian_mobile_numbers_supported": "Currently only Indian numbers are supported",
   "op_encounter": "OP Encounter",
@@ -1712,6 +1715,7 @@
   "register_patient": "Register Patient",
   "reject": "Reject",
   "rejected": "Rejected",
+  "related_person": "Related Person",
   "reload": "Reload",
   "remarks": "Remarks",
   "remarks_placeholder": "Enter remarks",
@@ -1972,6 +1976,7 @@
   "status": "Status",
   "stop": "Stop",
   "stop_recording": "Stop Recording",
+  "stopped": "Stopped",
   "stream_stop_due_to_inativity": "The live feed will stop streaming due to inactivity",
   "stream_stopped_due_to_inativity": "The live feed has stopped streaming due to inactivity",
   "stream_uuid": "Stream UUID",
