+++ src/components/Questionnaire/QuestionTypes/MedicationRequestQuestion.tsx
@@ -1,4 +1,4 @@
-import { MinusCircledIcon } from "@radix-ui/react-icons";
+import { MinusCircledIcon, Pencil2Icon } from "@radix-ui/react-icons";
 import { t } from "i18next";
 import React, { useState } from "react";
 
@@ -227,7 +227,7 @@ export function MedicationRequestQuestion({
                           className={cn(
                             "flex items-center gap-2 px-2 py-0.5 rounded-md shadow-sm text-sm",
                             expandedMedicationIndex === index
-                              ? "bg-white"
+                              ? "bg-gray-50"
                               : "bg-gray-100",
                           )}
                         >
@@ -239,6 +239,7 @@ export function MedicationRequestQuestion({
                           <div className="flex items-center gap-1">
                             {expandedMedicationIndex !== index && (
                               <Button
+                                aria-label="Expand Medication Request"
                                 variant="ghost"
                                 size="icon"
                                 className="h-8 w-8 text-gray-500 hover:text-gray-900"
@@ -248,21 +249,7 @@ export function MedicationRequestQuestion({
                                 }}
                                 disabled={disabled}
                               >
-                                <svg
-                                  width="15"
-                                  height="15"
-                                  viewBox="0 0 15 15"
-                                  fill="none"
-                                  xmlns="http://www.w3.org/2000/svg"
-                                  className="h-4 w-4"
-                                >
-                                  <path
-                                    d="M12.1464 1.14645C12.3417 0.951184 12.6583 0.951184 12.8535 1.14645L14.8535 3.14645C15.0488 3.34171 15.0488 3.65829 14.8535 3.85355L10.9109 7.79618C10.8349 7.87218 10.7471 7.93543 10.651 7.9835L6.72359 9.94721C6.53109 10.0435 6.29861 10.0057 6.14643 9.85355C5.99425 9.70137 5.95652 9.46889 6.05277 9.27639L8.01648 5.34897C8.06455 5.25283 8.1278 5.16507 8.2038 5.08907L12.1464 1.14645ZM12.5 2.20711L8.91091 5.79618L7.87266 7.87267L8.12731 8.12732L10.2038 7.08907L13.7929 3.5L12.5 2.20711ZM9.99998 2L8.99998 3H4.9C4.47171 3 4.18056 3.00039 3.95552 3.01877C3.73631 3.03668 3.62421 3.06915 3.54601 3.10899C3.35785 3.20487 3.20487 3.35785 3.10899 3.54601C3.06915 3.62421 3.03669 3.73631 3.01878 3.95552C3.00039 4.18056 3 4.47171 3 4.9V11.1C3 11.5283 3.00039 11.8194 3.01878 12.0445C3.03669 12.2637 3.06915 12.3758 3.10899 12.454C3.20487 12.6422 3.35785 12.7951 3.54601 12.891C3.62421 12.9309 3.73631 12.9633 3.95552 12.9812C4.18056 12.9996 4.47171 13 4.9 13H11.1C11.5283 13 11.8194 12.9996 12.0445 12.9812C12.2637 12.9633 12.3758 12.9309 12.454 12.891C12.6422 12.7951 12.7951 12.6422 12.891 12.454C12.9309 12.3758 12.9633 12.2637 12.9812 12.0445C12.9996 11.8194 13 11.5283 13 11.1V7.00002L12 8.00002V11.1V11.1207C12 11.5231 11.9995 11.7666 11.9862 11.9402C11.9735 12.1033 11.9514 12.1558 11.9397 12.1775C11.9092 12.2321 11.8321 12.3092 11.7775 12.3397C11.7557 12.3514 11.7033 12.3735 11.5402 12.3862C11.3666 12.3995 11.1231 12.4 10.7207 12.4H5.27924C4.87691 12.4 4.63338 12.3995 4.45981 12.3862C4.29672 12.3735 4.24424 12.3514 4.22252 12.3397C4.16792 12.3092 4.09077 12.2321 4.06028 12.1775C4.04858 12.1558 4.02644 12.1033 4.01375 11.9402C4.00045 11.7666 4 11.5231 4 11.1207V4.87924C4 4.47691 4.00045 4.23338 4.01375 4.05981C4.02644 3.89672 4.04858 3.84424 4.06028 3.82252C4.09077 3.76792 4.16792 3.69077 4.22252 3.66028C4.24424 3.64858 4.29672 3.62644 4.45981 3.61375C4.63338 3.60045 4.87691 3.6 5.27924 3.6H8.99998L9.99998 2Z"
-                                    fill="currentColor"
-                                    fillRule="evenodd"
-                                    clipRule="evenodd"
-                                  ></path>
-                                </svg>
+                                <Pencil2Icon className="h-4 w-4" />
                               </Button>
                             )}
                             <Button
